+++ src/etcd/auth.py
@@ -1,7 +1,8 @@
 import json
 
 import logging
-import etcd
+
+from .commom import EtcdInsufficientPermissions, EtcdKeyNotFound, EtcdException
 
 _log = logging.getLogger(__name__)
 
@@ -25,17 +26,17 @@ class EtcdAuthBase(object):
     def read(self):
         try:
             response = self.client.api_execute(self.uri, self.client._MGET)
-        except etcd.EtcdInsufficientPermissions as e:
+        except EtcdInsufficientPermissions as e:
             _log.error("Any action on the authorization requires the root role")
             raise
-        except etcd.EtcdKeyNotFound:
+        except EtcdKeyNotFound:
             _log.info("%s '%s' not found", self.entity, self.name)
             raise
         except Exception as e:
             _log.error("Failed to fetch %s in %s%s: %r",
                        self.entity, self.client._base_uri,
                        self.client.version_prefix, e)
-            raise etcd.EtcdException(
+            raise EtcdException(
                 "Could not fetch {} '{}'".format(self.entity, self.name))
 
         self._from_net(response.data)
@@ -44,7 +45,7 @@ class EtcdAuthBase(object):
         try:
             r = self.__class__(self.client, self.name)
             r.read()
-        except etcd.EtcdKeyNotFound:
+        except EtcdKeyNotFound:
             r = None
         try:
             for payload in self._to_net(r):
@@ -53,29 +54,29 @@ class EtcdAuthBase(object):
                                                         params=payload)
                 # This will fail if the response is an error
                 self._from_net(response.data)
-        except etcd.EtcdInsufficientPermissions as e:
+        except EtcdInsufficientPermissions as e:
             _log.error("Any action on the authorization requires the root role")
             raise
         except Exception as e:
             _log.error("Failed to write %s '%s'", self.entity, self.name)
             # TODO: fine-grained exception handling
-            raise etcd.EtcdException(
+            raise EtcdException(
                 "Could not write {} '{}': {}".format(self.entity,
                                                      self.name, e))
 
     def delete(self):
         try:
             _ = self.client.api_execute(self.uri, self.client._MDELETE)
-        except etcd.EtcdInsufficientPermissions as e:
+        except EtcdInsufficientPermissions as e:
             _log.error("Any action on the authorization requires the root role")
             raise
-        except etcd.EtcdKeyNotFound:
+        except EtcdKeyNotFound:
             _log.info("%s '%s' not found", self.entity, self.name)
             raise
         except Exception as e:
             _log.error("Failed to delete %s in %s%s: %r",
                        self.entity, self._base_uri, self.version_prefix, e)
-            raise etcd.EtcdException(
+            raise EtcdException(
                 "Could not delete {} '{}'".format(self.entity, self.name))
 
     def _from_net(self, data):
@@ -92,7 +93,7 @@ class EtcdAuthBase(object):
 
 
 class EtcdUser(EtcdAuthBase):
-    """Class to manage in a orm-like way etcd users"""
+    """Class to manage in a orm-like way users"""
     entity = 'user'
 
     def __init__(self, client, name):
