+++ src/etcd/tests/test_auth.py
@@ -1,6 +1,7 @@
-from etcd.tests.integration.test_simple import EtcdIntegrationTest
-from etcd import auth
-import etcd
+from .integration.test_simple import EtcdIntegrationTest
+from ..auth import EtcdUser, Auth, EtcdRole
+from ..client import Client
+from ..commom import EtcdKeyNotFound, EtcdInsufficientPermissions, EtcdException
 
 
 class TestEtcdAuthBase(EtcdIntegrationTest):
@@ -8,18 +9,18 @@ class TestEtcdAuthBase(EtcdIntegrationTest):
 
     def setUp(self):
         # Sets up the root user, toggles auth
-        u = auth.EtcdUser(self.client, 'root')
+        u = EtcdUser(self.client, 'root')
         u.password = 'testpass'
         u.write()
-        self.client = etcd.Client(port=6001, username='root',
+        self.client = Client(port=6001, username='root',
                                 password='testpass')
-        self.unauth_client = etcd.Client(port=6001)
-        a = auth.Auth(self.client)
+        self.unclient = Client(port=6001)
+        a = Auth(self.client)
         a.active = True
 
     def tearDown(self):
-        u = auth.EtcdUser(self.client, 'test_user')
-        r = auth.EtcdRole(self.client, 'test_role')
+        u = EtcdUser(self.client, 'test_user')
+        r = EtcdRole(self.client, 'test_role')
         try:
             u.delete()
         except:
@@ -28,17 +29,17 @@ class TestEtcdAuthBase(EtcdIntegrationTest):
             r.delete()
         except:
             pass
-        a = auth.Auth(self.client)
+        a = Auth(self.client)
         a.active = False
 
 
 class EtcdUserTest(TestEtcdAuthBase):
     def test_names(self):
-        u = auth.EtcdUser(self.client, 'test_user')
+        u = EtcdUser(self.client, 'test_user')
         self.assertEquals(u.names, ['root'])
 
     def test_read(self):
-        u = auth.EtcdUser(self.client, 'root')
+        u = EtcdUser(self.client, 'root')
         # Reading an existing user succeeds
         try:
             u.read()
@@ -53,21 +54,21 @@ class EtcdUserTest(TestEtcdAuthBase):
                                          'roles': ['root']}])
 
         # An inexistent user raises the appropriate exception
-        u = auth.EtcdUser(self.client, 'user.does.not.exist')
-        self.assertRaises(etcd.EtcdKeyNotFound, u.read)
+        u = EtcdUser(self.client, 'user.does.not.exist')
+        self.assertRaises(EtcdKeyNotFound, u.read)
 
-        # Reading with an unauthenticated client raises an exception
-        u = auth.EtcdUser(self.unauth_client, 'root')
-        self.assertRaises(etcd.EtcdInsufficientPermissions, u.read)
+        # Reading with an unnticated client raises an exception
+        u = EtcdUser(self.unclient, 'root')
+        self.assertRaises(EtcdInsufficientPermissions, u.read)
 
         # Generic errors are caught
-        c = etcd.Client(port=9999)
-        u = auth.EtcdUser(c, 'root')
-        self.assertRaises(etcd.EtcdException, u.read)
+        c = Client(port=9999)
+        u = EtcdUser(c, 'root')
+        self.assertRaises(EtcdException, u.read)
 
     def test_write_and_delete(self):
         # Create an user
-        u = auth.EtcdUser(self.client, 'test_user')
+        u = EtcdUser(self.client, 'test_user')
         u.roles.add('guest')
         u.roles.add('root')
         # directly from my suitcase
@@ -79,15 +80,15 @@ class EtcdUserTest(TestEtcdAuthBase):
         # Password gets wiped
         self.assertEquals(u.password, None)
         u.read()
-        # Verify we can log in as this user and access the auth (it has the
+        # Verify we can log in as this user and access the (it has the
         # root role)
-        cl = etcd.Client(port=6001, username='test_user',
+        cl = Client(port=6001, username='test_user',
                          password='123456')
-        ul = auth.EtcdUser(cl, 'root')
+        ul = EtcdUser(cl, 'root')
         try:
             ul.read()
-        except etcd.EtcdInsufficientPermissions:
-            self.fail("Reading auth with the new user is not possible")
+        except EtcdInsufficientPermissions:
+            self.fail("Reading with the new user is not possible")
 
         self.assertEquals(u.name, "test_user")
         self.assertEquals(u.roles, set(['guest', 'root']))
@@ -100,28 +101,28 @@ class EtcdUserTest(TestEtcdAuthBase):
         u.read()
         self.assertIn('test_group', u.roles)
 
-        # Unauthorized access is properly handled
-        ua = auth.EtcdUser(self.unauth_client, 'test_user')
-        self.assertRaises(etcd.EtcdInsufficientPermissions, ua.write)
+        # Unrized access is properly handled
+        ua = EtcdUser(self.unclient, 'test_user')
+        self.assertRaises(EtcdInsufficientPermissions, ua.write)
 
         # now let's test deletion
-        du = auth.EtcdUser(self.client, 'user.does.not.exist')
-        self.assertRaises(etcd.EtcdKeyNotFound, du.delete)
+        du = EtcdUser(self.client, 'user.does.not.exist')
+        self.assertRaises(EtcdKeyNotFound, du.delete)
 
         # Delete test_user
         u.delete()
-        self.assertRaises(etcd.EtcdKeyNotFound, u.read)
+        self.assertRaises(EtcdKeyNotFound, u.read)
         # Permissions are properly handled
-        self.assertRaises(etcd.EtcdInsufficientPermissions, ua.delete)
+        self.assertRaises(EtcdInsufficientPermissions, ua.delete)
 
 
 class EtcdRoleTest(TestEtcdAuthBase):
     def test_names(self):
-        r = auth.EtcdRole(self.client, 'guest')
+        r = EtcdRole(self.client, 'guest')
         self.assertListEqual(r.names, [u'guest', u'root'])
 
     def test_read(self):
-        r = auth.EtcdRole(self.client, 'guest')
+        r = EtcdRole(self.client, 'guest')
         try:
             r.read()
         except:
@@ -132,14 +133,14 @@ class EtcdRoleTest(TestEtcdAuthBase):
         # with EtcdUser
 
     def test_write_and_delete(self):
-        r = auth.EtcdRole(self.client, 'test_role')
+        r = EtcdRole(self.client, 'test_role')
         r.acls = {'*': 'R', '/test/*': 'RW'}
         try:
             r.write()
         except:
             self.fail("Writing a simple groups should not fail")
 
-        r1 = auth.EtcdRole(self.client, 'test_role')
+        r1 = EtcdRole(self.client, 'test_role')
         r1.read()
         self.assertEquals(r1.acls, r.acls)
         r.revoke('/test/*', 'W')
@@ -152,10 +153,10 @@ class EtcdRoleTest(TestEtcdAuthBase):
         self.assertEquals(r1.acls['/pub/*'], 'RW')
         # All other exceptions are tested by the user tests
         r1.name = None
-        self.assertRaises(etcd.EtcdException, r1.write)
+        self.assertRaises(EtcdException, r1.write)
         # ditto for delete
         try:
             r.delete()
         except:
             self.fail("A normal delete should not fail")
-        self.assertRaises(etcd.EtcdKeyNotFound, r.read)
+        self.assertRaises(EtcdKeyNotFound, r.read)
