+++ src/etcd/tests/unit/test_old_request.py
@@ -1,11 +1,13 @@
-import etcd
 import unittest
+
+from ...client import Client
+
 try:
     import mock
 except ImportError:
     from unittest import mock
 
-from etcd import EtcdException
+from ...commom import EtcdException, EtcdResult, EtcdKeyNotFound, EtcdNotFile
 
 
 class FakeHTTPResponse(object):
@@ -28,7 +30,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_set(self):
         """ Can set a value """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(201,
                                           '{"action":"SET",'
@@ -43,7 +45,7 @@ class TestClientRequest(unittest.TestCase):
         result = client.set('/testkey', 'test', ttl=19)
 
         self.assertEquals(
-            etcd.EtcdResult(
+            EtcdResult(
                 **{u'action': u'SET',
                    'node': {
                        u'expiration': u'2013-09-14T00:56:59.316195568+02:00',
@@ -55,7 +57,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_test_and_set(self):
         """ Can test and set a value """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"SET",'
@@ -68,7 +70,7 @@ class TestClientRequest(unittest.TestCase):
         )
         result = client.test_and_set('/testkey', 'newvalue', 'test', ttl=19)
         self.assertEquals(
-            etcd.EtcdResult(
+            EtcdResult(
                 **{u'action': u'SET',
                    u'node': {
                        u'expiration': u'2013-09-14T02:09:44.24390976+02:00',
@@ -82,7 +84,7 @@ class TestClientRequest(unittest.TestCase):
     def test_test_and_test_failure(self):
         """ Exception will be raised if prevValue != value in test_set """
 
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             side_effect=ValueError(
                 'The given PrevValue is not equal'
@@ -100,7 +102,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_delete(self):
         """ Can delete a value """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"DELETE",'
@@ -111,7 +113,7 @@ class TestClientRequest(unittest.TestCase):
                                           '"modifiedIndex":189}}')
         )
         result = client.delete('/testkey')
-        self.assertEquals(etcd.EtcdResult(
+        self.assertEquals(EtcdResult(
             **{u'action': u'DELETE',
                u'node': {
                    u'expiration': u'2013-09-14T01:06:35.5242587+02:00',
@@ -122,7 +124,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_get(self):
         """ Can get a value """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"GET",'
@@ -133,7 +135,7 @@ class TestClientRequest(unittest.TestCase):
         )
 
         result = client.get('/testkey')
-        self.assertEquals(etcd.EtcdResult(
+        self.assertEquals(EtcdResult(
             **{u'action': u'GET',
                u'node': {
                    u'modifiedIndex': 190,
@@ -151,14 +153,14 @@ class TestClientRequest(unittest.TestCase):
 
     def test_not_in(self):
         """ Can check if key is not in client """
-        client = etcd.Client()
-        client.get = mock.Mock(side_effect=etcd.EtcdKeyNotFound())
+        client = Client()
+        client.get = mock.Mock(side_effect=EtcdKeyNotFound())
         result = '/testkey' not in client
         self.assertEquals(True, result)
 
     def test_in(self):
         """ Can check if key is in client """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"GET",'
@@ -173,7 +175,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_simple_watch(self):
         """ Can watch values """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"SET",'
@@ -187,7 +189,7 @@ class TestClientRequest(unittest.TestCase):
         )
         result = client.watch('/testkey')
         self.assertEquals(
-            etcd.EtcdResult(
+            EtcdResult(
                 **{u'action': u'SET',
                    u'node': {
                        u'expiration': u'2013-09-14T01:35:07.623681365+02:00',
@@ -200,7 +202,7 @@ class TestClientRequest(unittest.TestCase):
 
     def test_index_watch(self):
         """ Can watch values from index """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"SET",'
@@ -214,7 +216,7 @@ class TestClientRequest(unittest.TestCase):
         )
         result = client.watch('/testkey', index=180)
         self.assertEquals(
-            etcd.EtcdResult(
+            EtcdResult(
                 **{u'action': u'SET',
                    u'node': {
                        u'expiration': u'2013-09-14T01:35:07.623681365+02:00',
@@ -229,7 +231,7 @@ class TestClientRequest(unittest.TestCase):
 class TestEventGenerator(object):
 
     def check_watch(self, result):
-        assert etcd.EtcdResult(
+        assert EtcdResult(
             **{u'action': u'SET',
                u'node': {
                    u'expiration': u'2013-09-14T01:35:07.623681365+02:00',
@@ -242,7 +244,7 @@ class TestEventGenerator(object):
 
     def test_eternal_watch(self):
         """ Can watch values from generator """
-        client = etcd.Client()
+        client = Client()
         client.api_execute = mock.Mock(
             return_value=FakeHTTPResponse(200,
                                           '{"action":"SET",'
@@ -263,7 +265,7 @@ class TestClientApiExecutor(unittest.TestCase):
 
     def test_get(self):
         """ http get request """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=200, data='arbitrary json data')
         client.http.request = mock.Mock(return_value=response)
         result = client.api_execute('/v1/keys/testkey', client._MGET)
@@ -271,7 +273,7 @@ class TestClientApiExecutor(unittest.TestCase):
 
     def test_delete(self):
         """ http delete request """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=200, data='arbitrary json data')
         client.http.request = mock.Mock(return_value=response)
         result = client.api_execute('/v1/keys/testkey', client._MDELETE)
@@ -279,7 +281,7 @@ class TestClientApiExecutor(unittest.TestCase):
 
     def test_get_error(self):
         """ http get error request 101"""
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=400,
                                     data='{"message": "message",'
                                     ' "cause": "cause",'
@@ -288,12 +290,12 @@ class TestClientApiExecutor(unittest.TestCase):
         try:
             client.api_execute('/v2/keys/testkey', client._MGET)
             assert False
-        except etcd.EtcdKeyNotFound as e:
+        except EtcdKeyNotFound as e:
             self.assertEquals(str(e), 'message : cause')
 
     def test_put(self):
         """ http put request """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=200, data='arbitrary json data')
         client.http.request_encode_body = mock.Mock(return_value=response)
         result = client.api_execute('/v2/keys/testkey', client._MPUT)
@@ -301,7 +303,7 @@ class TestClientApiExecutor(unittest.TestCase):
 
     def test_test_and_set_error(self):
         """ http post error request 101 """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(
             status=400,
             data='{"message": "message", "cause": "cause", "errorCode": 101}')
@@ -315,7 +317,7 @@ class TestClientApiExecutor(unittest.TestCase):
 
     def test_set_not_file_error(self):
         """ http post error request 102 """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(
             status=400,
             data='{"message": "message", "cause": "cause", "errorCode": 102}')
@@ -324,12 +326,12 @@ class TestClientApiExecutor(unittest.TestCase):
         try:
             client.api_execute('/v2/keys/testkey', client._MPUT, payload)
             self.fail()
-        except etcd.EtcdNotFile as e:
+        except EtcdNotFile as e:
             self.assertEquals('message : cause', str(e))
 
     def test_get_error_unknown(self):
         """ http get error request unknown """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=400,
                                     data='{"message": "message",'
                                          ' "cause": "cause",'
@@ -338,27 +340,27 @@ class TestClientApiExecutor(unittest.TestCase):
         try:
             client.api_execute('/v2/keys/testkey', client._MGET)
             self.fail()
-        except etcd.EtcdException as e:
+        except EtcdException as e:
             self.assertEqual(str(e), "message : cause")
 
     def test_get_error_request_invalid(self):
         """ http get error request invalid """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=400,
                                     data='{)*garbage')
         client.http.request = mock.Mock(return_value=response)
         try:
             client.api_execute('/v2/keys/testkey', client._MGET)
             self.fail()
-        except etcd.EtcdException as e:
+        except EtcdException as e:
             self.assertEqual(str(e),
                              "Bad response : {)*garbage")
 
     def test_get_error_invalid(self):
         """ http get error request invalid """
-        client = etcd.Client()
+        client = Client()
         response = FakeHTTPResponse(status=400,
                                     data='{){){)*garbage*')
         client.http.request = mock.Mock(return_value=response)
-        self.assertRaises(etcd.EtcdException, client.api_execute,
+        self.assertRaises(EtcdException, client.api_execute,
                           '/v2/keys/testkey', client._MGET)
