+++ src/etcd/tests/unit/test_request.py
@@ -1,8 +1,14 @@
 import socket
 import urllib3
 
-import etcd
-from etcd.tests.unit import TestClientApiBase
+from . import TestClientApiBase
+from ...client import Client
+from ...commom import EtcdResult, EtcdException, EtcdKeyNotFound,\
+    EtcdConnectionFailed, EtcdClusterIdChanged, EtcdWatchTimedOut
+
+
+
+
 
 try:
     import mock
@@ -88,13 +94,13 @@ class TestClientApiInterface(TestClientApiBase):
         mocker.return_value = self._prepare_response(200, d)
         self.assertEquals(data, self.client.machines)
 
-    @mock.patch('etcd.Client.machines', new_callable=mock.PropertyMock)
+    @mock.patch('Client.machines', new_callable=mock.PropertyMock)
     def test_use_proxies(self, mocker):
         """Do not overwrite the machines cache when using proxies"""
         mocker.return_value = ['https://10.0.0.2:4001',
                                'https://10.0.0.3:4001',
                                'https://10.0.0.4:4001']
-        c = etcd.Client(
+        c = Client(
             host=(('localhost', 4001), ('localproxy', 4001)),
             protocol='https',
             allow_reconnect=True,
@@ -105,7 +111,7 @@ class TestClientApiInterface(TestClientApiBase):
         self.assertEquals(c._base_uri, 'https://localhost:4001')
         self.assertNotIn(c.base_uri, c._machines_cache)
 
-        c = etcd.Client(
+        c = Client(
             host=(('localhost', 4001), ('10.0.0.2', 4001)),
             protocol='https',
             allow_reconnect=True,
@@ -157,7 +163,7 @@ class TestClientApiInterface(TestClientApiBase):
         self.assertEquals(self.client.leader_stats['leader'], "924e2e83e93f2560")
 
 
-    @mock.patch('etcd.Client.members', new_callable=mock.PropertyMock)
+    @mock.patch('Client.members', new_callable=mock.PropertyMock)
     def test_leader(self, mocker):
         """ Can request the leader """
         members = {"ce2a822cea30bfca": {"id": "ce2a822cea30bfca", "name": "default"}}
@@ -179,7 +185,7 @@ class TestClientApiInterface(TestClientApiBase):
 
         self._mock_api(200, d)
         res = self.client.write('/testkey', 'test')
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_update(self):
         """Can update a result."""
@@ -215,14 +221,14 @@ class TestClientApiInterface(TestClientApiBase):
         self._mock_api(201, d)
         res = self.client.write('/testkey', 'test')
         d['node']['newKey'] = True
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
 
 
     def test_not_found_response(self):
         """ Can handle server not found response """
         self._mock_api(404, 'Not found')
-        self.assertRaises(etcd.EtcdException, self.client.read, '/somebadkey')
+        self.assertRaises(EtcdException, self.client.read, '/somebadkey')
 
     def test_compare_and_swap(self):
         """ Can set compare-and-swap a value """
@@ -238,7 +244,7 @@ class TestClientApiInterface(TestClientApiBase):
 
         self._mock_api(200, d)
         res = self.client.write('/testkey', 'test', prevValue='test_old')
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_compare_and_swap_failure(self):
         """ Exception will be raised if prevValue != value in test_set """
@@ -268,7 +274,7 @@ class TestClientApiInterface(TestClientApiBase):
 
     def test_set_dir_with_value(self):
         """ Creating a directory with a value raises an error. """
-        self.assertRaises(etcd.EtcdException, self.client.write,
+        self.assertRaises(EtcdException, self.client.write,
                           '/bar', 'testvalye', dir=True)
 
     def test_delete(self):
@@ -283,7 +289,7 @@ class TestClientApiInterface(TestClientApiBase):
         }
         self._mock_api(200, d)
         res = self.client.delete('/testKey')
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_pop(self):
         """ Can pop a value """
@@ -302,7 +308,7 @@ class TestClientApiInterface(TestClientApiBase):
         self._mock_api(200, d)
         res = self.client.pop(d['node']['key'])
         self.assertEquals({attr: getattr(res, attr) for attr in dir(res)
-                           if attr in etcd.EtcdResult._node_props}, d['prevNode'])
+                           if attr in EtcdResult._node_props}, d['prevNode'])
         self.assertEqual(res.value, d['prevNode']['value'])
 
     def test_read(self):
@@ -317,7 +323,7 @@ class TestClientApiInterface(TestClientApiBase):
         }
         self._mock_api(200, d)
         res = self.client.read('/testKey')
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_get_dir(self):
         """Can get values in dirs"""
@@ -343,11 +349,11 @@ class TestClientApiInterface(TestClientApiBase):
         }
         self._mock_api(200, d)
         res = self.client.read('/testDir', recursive=True)
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_not_in(self):
         """ Can check if key is not in client """
-        self._mock_exception(etcd.EtcdKeyNotFound, 'Key not Found : /testKey')
+        self._mock_exception(EtcdKeyNotFound, 'Key not Found : /testKey')
         self.assertTrue('/testey' not in self.client)
 
     def test_in(self):
@@ -375,7 +381,7 @@ class TestClientApiInterface(TestClientApiBase):
         }
         self._mock_api(200, d)
         res = self.client.read('/testkey', wait=True)
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
     def test_watch_index(self):
         """ Can watch a key starting from the given Index """
@@ -389,13 +395,13 @@ class TestClientApiInterface(TestClientApiBase):
         }
         self._mock_api(200, d)
         res = self.client.read('/testkey', wait=True, waitIndex=True)
-        self.assertEquals(res, etcd.EtcdResult(**d))
+        self.assertEquals(res, EtcdResult(**d))
 
 
 class TestClientRequest(TestClientApiInterface):
 
     def setUp(self):
-        self.client = etcd.Client(expected_cluster_id="abcdef1234")
+        self.client = Client(expected_cluster_id="abcdef1234")
 
     def _mock_api(self, status, d, cluster_id=None):
         resp = self._prepare_response(status, d)
@@ -439,7 +445,7 @@ class TestClientRequest(TestClientApiInterface):
                                                             "Read timed out")
         )
         self.assertRaises(
-            etcd.EtcdWatchTimedOut,
+            EtcdWatchTimedOut,
             self.client.watch,
             '/testKey',
         )
@@ -451,7 +457,7 @@ class TestClientRequest(TestClientApiInterface):
 
     def test_api_method_not_supported(self):
         """ Exception will be raised if an unsupported HTTP method is used """
-        self.assertRaises(etcd.EtcdException,
+        self.assertRaises(EtcdException,
                           self.client.api_execute, '/testpath/bar', 'TRACE')
 
     def test_read_cluster_id_changed(self):
@@ -467,7 +473,7 @@ class TestClientRequest(TestClientApiInterface):
             }
         }
         self._mock_api(200, d, cluster_id="notabcd1234")
-        self.assertRaises(etcd.EtcdClusterIdChanged,
+        self.assertRaises(EtcdClusterIdChanged,
                           self.client.read, '/testkey')
         self.client.read("/testkey")
 
@@ -476,10 +482,10 @@ class TestClientRequest(TestClientApiInterface):
             self.client.http.request,
             side_effect=socket.error()
         )
-        self.assertRaises(etcd.EtcdConnectionFailed,
+        self.assertRaises(EtcdConnectionFailed,
                           self.client.read, '/something')
         # Direct GET request
-        self.assertRaises(etcd.EtcdConnectionFailed,
+        self.assertRaises(EtcdConnectionFailed,
                           self.client.api_execute, '/a', 'GET')
 
     def test_not_in(self):
@@ -500,7 +506,7 @@ class TestClientRequest(TestClientApiInterface):
                 u'value': u'test'
             }
         }
-        res = etcd.EtcdResult(**d)
+        res = EtcdResult(**d)
 
         error = {
             "errorCode": 101,
