+++ src/etcd/tests/unit/test_result.py
@@ -1,7 +1,5 @@
-import etcd
 import unittest
-import json
-import urllib3
+from ...commom import EtcdResult
 
 try:
     import mock
@@ -24,7 +22,7 @@ class TestEtcdResult(unittest.TestCase):
             'newKey': False,
             'dir': False,
         }}
-        result = etcd.EtcdResult(**response)
+        result = EtcdResult(**response)
         self.assertEqual(result.key, response["node"]["key"])
         self.assertEqual(result.value, response["node"]["value"])
 
@@ -68,7 +66,7 @@ class TestEtcdResult(unittest.TestCase):
             'dir': True,
             'nodes': [leaf0, leaf1]
         }}
-        result = etcd.EtcdResult(**testnode)
+        result = EtcdResult(**testnode)
         self.assertEqual(result.key, "/test/")
         self.assertTrue(result.dir)
 
@@ -120,7 +118,7 @@ class TestEtcdResult(unittest.TestCase):
             'dir': True,
             'nodes': [mid0, mid1]
         }}
-        result = etcd.EtcdResult(**testnode)
+        result = EtcdResult(**testnode)
         self.assertEqual(result.key, "/test/")
         self.assertTrue(result.dir)
